refactor(test): use shared IRewriteRuleRepository mock
Replace local createMockRepository.ts with import from shared location
at tests/unit/application/ports/IRewriteRuleRepository/. This follows
the project's mock guidelines to avoid duplication.
