--- host-frontend-root/frontend-src-root/tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/mocks/createMockRepository.ts
@@ -1,16 +1,0 @@
-import { vi } from 'vitest';
-
-import { IRewriteRuleRepository } from 'src/application-business-rules/ports/gateway/IRewriteRuleRepository';
-
-/**
- * IRewriteRuleRepositoryのモックオブジェクトを生成する
- * @returns IRewriteRuleRepository型のモックオブジェクト
- */
-export const createMockRepository = (): IRewriteRuleRepository => ({
-  create: vi.fn(),
-  update: vi.fn().mockResolvedValue(undefined),
-  delete: vi.fn().mockResolvedValue(undefined),
-  getAll: vi.fn(),
-  getById: vi.fn(),
-  getRulesMatchingUrl: vi.fn(),
-});
